Create oneAlbum page, links on allAlbums, fix url
I fixed the url routing so that slugs are handed off to
the views. I created a getAbsoluteURL function to make the
links on the allAlbums page work for a given obj. I made it
so that the images for an obj render on the oneAlbum page.
I created a CSS file for the oneAlbum page as well.